requirements.txt
folium==0.14.0
geopy==2.3.0
inflection==0.5.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
Pillow==10.0.0
plotly==5.15.0
streamlit==1.24.0
streamlit_folium==0.12.0
